Cleaned up some tests, added some conveinence API's
