+++ Tests/Singularity.Duality.Test/Singularity.Duality.Test.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFrameworks>net472;netcoreapp2.0</TargetFrameworks>
     <Configurations>Debug;Release;z_CI_config</Configurations>
   </PropertyGroup>
