+++ Tests/Singularity.Test/ContainerTests.cs
@@ -9,294 +9,6 @@ namespace Singularity.Test
 	{
 		public class Nested
 		{
-			public class Decorators
-			{
-                [Fact]
-                public void GetInstance_DecoratorInRoot_1Deep_DecoratorsAreCorrectlyApplied()
-                {
-                    var rootConfig = new BindingConfig();
-                    rootConfig.Decorate<IComponent, Decorator1>();
-
-                    using (var rootContainer = new Container(rootConfig))
-                    {
-                        var nested1Config = new BindingConfig();
-                        nested1Config.Register<IComponent, Component>();
-                        using (Container nested1Container = rootContainer.GetNestedContainer(nested1Config))
-                        {
-                            var nested1Value = nested1Container.GetInstance<IComponent>();
-                            var nested1Decorator1 = Assert.IsType<Decorator1>(nested1Value);
-                            var component = Assert.IsType<Component>(nested1Decorator1.Component);
-                        }
-                    }
-                }
-
-                [Fact]
-			    public void GetInstance_DecoratorInRoot_2Deep_DecoratorsAreCorrectlyApplied()
-			    {
-			        var rootConfig = new BindingConfig();
-			        rootConfig.Decorate<IComponent, Decorator1>();
-
-			        using (var rootContainer = new Container(rootConfig))
-			        {
-			            var nested1Config = new BindingConfig();
-			            nested1Config.Decorate<IComponent, Decorator2>();
-                        using (Container nested1Container = rootContainer.GetNestedContainer(nested1Config))
-			            {
-			                var nested2Config = new BindingConfig();
-			                nested2Config.Register<IComponent, Component>();
-                            using (Container nested2Container = nested1Container.GetNestedContainer(nested2Config))
-			                {
-			                    var nested2Value = nested2Container.GetInstance<IComponent>();
-			                    var nested2Decorator2 = Assert.IsType<Decorator2>(nested2Value);
-                                var nested2Decorator1 = Assert.IsType<Decorator1>(nested2Decorator2.Component);
-                                var component = Assert.IsType<Component>(nested2Decorator1.Component);
-			                }
-			            }
-			        }
-			    }
-
-                [Fact]
-				public void GetInstance_DecoratorsAreCorrectlyApplied()
-				{
-					var config = new BindingConfig();
-					config.Decorate<IComponent, Decorator1>();
-					config.Register<IComponent, Component>();
-
-					using (var container = new Container(config))
-					{
-						var value = container.GetInstance<IComponent>();
-						Assert.NotNull(value);
-						Assert.Equal(typeof(Decorator1), value.GetType());
-						var decorator1 = (Decorator1)value;
-						Assert.Equal(typeof(Component), decorator1.Component.GetType());
-
-						var nestedConfig = new BindingConfig();
-						nestedConfig.Decorate<IComponent, Decorator2>();
-						using (Container nestedContainer = container.GetNestedContainer(nestedConfig))
-						{
-							var nestedValue = nestedContainer.GetInstance<IComponent>();
-							Assert.NotNull(nestedValue);
-							Assert.Equal(typeof(Decorator2), nestedValue.GetType());
-							var nestedDecorator2 = (Decorator2)nestedValue;
-							Assert.Equal(typeof(Decorator1), nestedDecorator2.Component.GetType());
-							var nestedDecorator1 = (Decorator1)nestedDecorator2.Component;
-
-							Assert.Equal(typeof(Component), nestedDecorator1.Component.GetType());
-						}
-					}
-				}
-
-			    [Fact]
-			    public void GetInstance_DecorateNestedContainer_Override_PerContainerLifetime()
-			    {
-			        var rootConfig = new BindingConfig();
-			        rootConfig.Decorate<IComponent, Decorator1>();
-			        rootConfig.Register<IComponent, Component>().With(CreationMode.Singleton);
-
-                    using (var rootContainer = new Container(rootConfig))
-                    {
-                        var rootValue = rootContainer.GetInstance<IComponent>();
-			            var nested1Config = new BindingConfig();
-			            nested1Config.Decorate<IComponent, Decorator2>();
-                        nested1Config.Register<IComponent, Component>().With(CreationMode.Singleton);
-                        using (Container nested1Container = rootContainer.GetNestedContainer(nested1Config))
-			            {
-			                var nested1Value = nested1Container.GetInstance<IComponent>();
-			                var nested1Decorator2 = Assert.IsType<Decorator2>(nested1Value);
-			                var nested1Decorator1 = Assert.IsType<Decorator1>(nested1Decorator2.Component);
-                            var component = Assert.IsType<Component>(nested1Decorator1.Component);
-			            }
-			        }
-                }
-
-                [Fact]
-				public void GetInstance_DecorateNestedContainer_PerContainerLifetime()
-				{
-					var config = new BindingConfig();
-
-					config.Decorate<ITestService11, TestService11_Decorator1>();
-
-					config.Register<ITestService10, TestService10>();
-					config.Register<ITestService11, TestService11>().With(CreationMode.Singleton);
-
-				    using (var container = new Container(config))
-				    {
-				        var value = container.GetInstance<ITestService11>();
-
-				        var decorator1 = Assert.IsType<TestService11_Decorator1>(value);
-				        var testService11 = Assert.IsType<TestService11>(decorator1.TestService11);
-				        Assert.IsType<TestService10>(testService11.TestService10);
-
-				        var nestedConfig = new BindingConfig();
-				        nestedConfig.Decorate<ITestService11, TestService11_Decorator2>();
-				        using (Container nestedContainer = container.GetNestedContainer(nestedConfig))
-				        {
-				            var nestedInstance = nestedContainer.GetInstance<ITestService11>();
-
-				            var nestedDecorator2 = Assert.IsType<TestService11_Decorator2>(nestedInstance);
-				            var nestedDecorator1 = Assert.IsType<TestService11_Decorator1>(nestedDecorator2.TestService11);
-				            Assert.IsType<TestService11>(nestedDecorator1.TestService11);
-				        }
-				    }
-				}
-
-				[Fact]
-				public void GetInstance_Decorate_PerContainerLifetime_SameInstanceIsReturned()
-				{
-					var config = new BindingConfig();
-					config.Register<IComponent, Component>().With(CreationMode.Singleton);
-
-					using (var container = new Container(config))
-					{
-						var value = container.GetInstance<IComponent>();
-					    Assert.IsType<Component>(value);
-
-						var nestedConfig = new BindingConfig();
-						nestedConfig.Decorate<IComponent, Decorator1>();
-						using (Container nestedContainer = container.GetNestedContainer(nestedConfig))
-						{
-							var nestedValue = nestedContainer.GetInstance<IComponent>();
-						    var decorator1 = Assert.IsType<Decorator1>(nestedValue);
-						    var component = Assert.IsType<Component>(decorator1.Component);
-                            Assert.Equal(value, component);
-						}
-					}
-				}
-			}
-
-			public class Disposed
-			{
-				[Fact]
-				public void GetInstance_PerContainerLifetimeAndOverride_IsDisposed()
-				{
-					var config = new BindingConfig();
-					config.Register<IDisposable, Disposable>().With(CreationMode.Singleton).OnDeath(x => x.Dispose());
-
-					var container = new Container(config);
-
-					var topLevelInstance = container.GetInstance<IDisposable>();
-					Assert.NotNull(topLevelInstance);
-					Assert.Equal(typeof(Disposable), topLevelInstance.GetType());
-
-					{
-						var nestedConfig = new BindingConfig();
-						nestedConfig.Register<IDisposable, Disposable>().With(CreationMode.Singleton).OnDeath(x => x.Dispose());
-						Container nestedContainer = container.GetNestedContainer(nestedConfig, new Scoped());
-						var nestedInstance = nestedContainer.GetInstance<IDisposable>();
-
-						Assert.NotNull(nestedInstance);
-						Assert.Equal(typeof(Disposable), nestedInstance.GetType());
-
-						var castednestedInstance = (Disposable)nestedInstance;
-						Assert.False(castednestedInstance.IsDisposed);
-						nestedContainer.Dispose();
-						Assert.True(castednestedInstance.IsDisposed);
-					}
-
-					var castedTopLevelInstance = (Disposable)topLevelInstance;
-					Assert.False(castedTopLevelInstance.IsDisposed);
-					container.Dispose();
-					Assert.True(castedTopLevelInstance.IsDisposed);
-				}
-
-				[Fact]
-				public void GetInstance_PerContainerLifetime_IsDisposedInTopLevel()
-				{
-					var config = new BindingConfig();
-					config.Register<IDisposable, Disposable>().With(CreationMode.Singleton).OnDeath(x => x.Dispose());
-
-					var container = new Container(config);
-
-					var topLevelInstance = container.GetInstance<IDisposable>();
-					Assert.NotNull(topLevelInstance);
-					Assert.Equal(typeof(Disposable), topLevelInstance.GetType());
-
-					{
-						var nestedConfig = new BindingConfig();
-
-						Container nestedContainer = container.GetNestedContainer(nestedConfig);
-						var nestedInstance = nestedContainer.GetInstance<IDisposable>();
-
-						Assert.NotNull(nestedInstance);
-						Assert.Equal(typeof(Disposable), nestedInstance.GetType());
-
-						var castednestedInstance = (Disposable)nestedInstance;
-						Assert.False(castednestedInstance.IsDisposed);
-						nestedContainer.Dispose();
-						Assert.False(castednestedInstance.IsDisposed);
-					}
-
-					var castedTopLevelInstance = (Disposable)topLevelInstance;
-					Assert.False(castedTopLevelInstance.IsDisposed);
-					container.Dispose();
-					Assert.True(castedTopLevelInstance.IsDisposed);
-				}
-
-				[Fact]
-				public void GetInstance_PerCallLifetime_IsDisposedInTopLevel()
-				{
-					var config = new BindingConfig();
-					config.Register<IDisposable, Disposable>().OnDeath(x => x.Dispose());
-
-					var container = new Container(config);
-
-					var topLevelInstance = container.GetInstance<IDisposable>();
-					Assert.NotNull(topLevelInstance);
-					Assert.Equal(typeof(Disposable), topLevelInstance.GetType());
-
-					{
-						var nestedConfig = new BindingConfig();
-
-						Container nestedContainer = container.GetNestedContainer(nestedConfig, new Scoped());
-						var nestedInstance = nestedContainer.GetInstance<IDisposable>();
-
-						Assert.NotNull(nestedInstance);
-						Assert.Equal(typeof(Disposable), nestedInstance.GetType());
-						Assert.NotEqual(nestedInstance, topLevelInstance);
-
-						var castednestedInstance = (Disposable)nestedInstance;
-						Assert.False(castednestedInstance.IsDisposed);
-						nestedContainer.Dispose();
-						Assert.True(castednestedInstance.IsDisposed);
-					}
-
-					var castedTopLevelInstance = (Disposable)topLevelInstance;
-					Assert.False(castedTopLevelInstance.IsDisposed);
-					container.Dispose();
-					Assert.True(castedTopLevelInstance.IsDisposed);
-				}
-
-				[Fact]
-				public void GetInstance_PerContainerLifetimeAndNestedContainerDecorator_IsDisposed()
-				{
-					var config = new BindingConfig();
-					config.Register<IDisposable, Disposable>().With(CreationMode.Singleton).OnDeath(x => x.Dispose());
-
-					var container = new Container(config);
-					var topLevelInstance = container.GetInstance<IDisposable>();
-
-					Assert.NotNull(topLevelInstance);
-					Assert.Equal(typeof(Disposable), topLevelInstance.GetType());
-					{
-						var nestedConfig = new BindingConfig();
-						nestedConfig.Decorate<IDisposable, DisposableDecorator>();
-						Container nestedContainer = container.GetNestedContainer(nestedConfig);
-						var nestedInstance = nestedContainer.GetInstance<IDisposable>();
-
-						Assert.NotNull(nestedInstance);
-						Assert.Equal(typeof(DisposableDecorator), nestedInstance.GetType());
-						var disposableDecorator = (DisposableDecorator)nestedInstance;
-						Assert.Equal(typeof(Disposable), disposableDecorator.Disposable.GetType());
-						var value = (Disposable)disposableDecorator.Disposable;
-
-						Assert.Equal(topLevelInstance, value);
-						Assert.False(value.IsDisposed);
-						container.Dispose();
-						Assert.True(value.IsDisposed);
-					}
-				}
-			}
-
 			public class WithDependencies
 			{
 				[Fact]
