+++ Tests/Singularity.Test/Injection/GenericsTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq.Expressions;
+﻿using System.Collections.Generic;
+using System.Linq.Expressions;
 using Singularity.TestClasses.TestClasses;
 using Xunit;
