+++ Tests/Singularity.Test/Singularity.Test.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFrameworks>net472;netcoreapp2.0</TargetFrameworks>
     <Configurations>Debug;Release;z_CI_config</Configurations>
     <NoWarn>1701;1702;1705;1591;NU1701;NU1702;CS1591;CS1574;CS1711</NoWarn>
     <LangVersion>8.0</LangVersion>
